requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
elastic-transport==8.15.1
elasticsearch==8.17.0
idna==3.10
piexif==1.1.3
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
watchdog==6.0.0
